refactor(gateway): move gatewayapi/common.go to package
