+++ pkg/plugins/runtime/k8s/controllers/gatewayapi/common/reconcile.go
@@ -1,4 +1,4 @@
-package gatewayapi
+package common
 
 import (
 	"context"
@@ -10,45 +10,19 @@ import (
 	kube_meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	kube_types "k8s.io/apimachinery/pkg/types"
 	kube_client "sigs.k8s.io/controller-runtime/pkg/client"
-	gatewayapi "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
-	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	k8s_model "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/pkg/model"
 	k8s_registry "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/pkg/registry"
 )
 
-const (
-	controllerName = gatewayapi.GatewayController("gateways.kuma.io/controller")
-	httpRouteKind  = gatewayapi.Kind("HTTPRoute")
-)
+const ownerLabel = "gateways.kuma.io/gateway.networking.k8s.io-owner"
 
-func getGatewayClass(ctx context.Context, client kube_client.Client, name gatewayapi.ObjectName) (*gatewayapi.GatewayClass, error) {
-	class := &gatewayapi.GatewayClass{}
-	classObjectKey := kube_types.NamespacedName{Name: string(name)}
-
-	if err := client.Get(ctx, classObjectKey, class); err != nil {
-		if kube_apierrs.IsNotFound(err) {
-			return nil, nil
-		}
-
-		return nil, errors.Wrapf(err, "failed to get GatewayClass %s", classObjectKey)
-	}
-
-	return class, nil
-}
-
-func serviceTagForGateway(name kube_types.NamespacedName) map[string]string {
-	return map[string]string{
-		mesh_proto.ServiceTag: fmt.Sprintf("%s_%s_gateway", name.Name, name.Namespace),
-	}
-}
-
-// reconcileLabelledObjectSet manages a set of owned kuma objects based on
+// ReconcileLabelledObject manages a set of owned kuma objects based on
 // labels with the owner key.
 // ownerMesh can be empty if the ownedSpec is nil.
 // ownedType tells us what type the owned object is.
 // ownedSpec should be set to nil if the object shouldn't exist.
-func reconcileLabelledObject(
+func ReconcileLabelledObject(
 	ctx context.Context,
 	registry k8s_registry.TypeRegistry,
 	client kube_client.Client,
