+++ pkg/plugins/runtime/k8s/controllers/gatewayapi/gateway_controller.go
@@ -21,6 +21,7 @@ import (
 	mesh_k8s "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/api/v1alpha1"
 	k8s_registry "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/pkg/registry"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/containers"
+	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/controllers/gatewayapi/common"
 	k8s_util "github.com/kumahq/kuma/pkg/plugins/runtime/k8s/util"
 )
 
@@ -44,19 +45,19 @@ func (r *GatewayReconciler) Reconcile(ctx context.Context, req kube_ctrl.Request
 		if kube_apierrs.IsNotFound(err) {
 			// We don't know the mesh, but we don't need it to delete our
 			// object.
-			err := reconcileLabelledObject(ctx, r.TypeRegistry, r.Client, req.NamespacedName, core_model.NoMesh, &mesh_proto.Gateway{}, nil)
+			err := common.ReconcileLabelledObject(ctx, r.TypeRegistry, r.Client, req.NamespacedName, core_model.NoMesh, &mesh_proto.Gateway{}, nil)
 			return kube_ctrl.Result{}, errors.Wrap(err, "could not delete owned Gateway.kuma.io")
 		}
 
 		return kube_ctrl.Result{}, err
 	}
 
-	class, err := getGatewayClass(ctx, r.Client, gateway.Spec.GatewayClassName)
+	class, err := common.GetGatewayClass(ctx, r.Client, gateway.Spec.GatewayClassName)
 	if err != nil {
 		return kube_ctrl.Result{}, errors.Wrap(err, "unable to retrieve GatewayClass referenced by Gateway")
 	}
 
-	if class.Spec.ControllerName != controllerName {
+	if class.Spec.ControllerName != common.ControllerName {
 		return kube_ctrl.Result{}, nil
 	}
 
@@ -72,7 +73,7 @@ func (r *GatewayReconciler) Reconcile(ctx context.Context, req kube_ctrl.Request
 
 	mesh := k8s_util.MeshOf(gateway, &ns)
 
-	if err := reconcileLabelledObject(ctx, r.TypeRegistry, r.Client, req.NamespacedName, mesh, &mesh_proto.Gateway{}, gatewaySpec); err != nil {
+	if err := common.ReconcileLabelledObject(ctx, r.TypeRegistry, r.Client, req.NamespacedName, mesh, &mesh_proto.Gateway{}, gatewaySpec); err != nil {
 		return kube_ctrl.Result{}, errors.Wrap(err, "could not reconcile owned Gateway.kuma.io")
 	}
 
@@ -99,7 +100,7 @@ func (r *GatewayReconciler) createOrUpdateInstance(ctx context.Context, gateway 
 	if _, err := kube_controllerutil.CreateOrUpdate(ctx, r.Client, instance, func() error {
 		instance.Spec = mesh_k8s.GatewayInstanceSpec{
 			ServiceType: kube_core.ServiceTypeLoadBalancer,
-			Tags:        serviceTagForGateway(kube_client.ObjectKeyFromObject(gateway)),
+			Tags:        common.ServiceTagForGateway(kube_client.ObjectKeyFromObject(gateway)),
 		}
 
 		err := kube_controllerutil.SetControllerReference(gateway, instance, r.Scheme)
